chore: moved memory_export adapter into src dir
